requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.3.0
pyserial==3.1.1
pytz==2016.7
Werkzeug==0.11.11
